Move console redirect into wandb_run, add first pass of wandb_reinit and join fixes (#77)
console and reinit work
